Setup: Fixed uninstallation of EP installations that have went through upgrades before the proper Pin to Start fix ^2

+++ ep_setup/ep_setup.c
@@ -985,11 +985,14 @@ int WINAPI wWinMain(
                 {
                     bOk = DeleteFileW(wszPath);
                 }
-                PathRemoveExtensionW(wszPath);
-                wcscat_s(wszPath, MAX_PATH, L".prev");
-                if (FileExistsW(wszPath))
+                if (bOk)
                 {
-                    bOk = DeleteFileW(wszPath);
+                    PathRemoveExtensionW(wszPath);
+                    wcscat_s(wszPath, MAX_PATH, L".prev");
+                    if (FileExistsW(wszPath))
+                    {
+                        bOk = DeleteFileW(wszPath);
+                    }
                 }
                 if (bOk)
                 {
